backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
celery==5.3.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
apify-client==1.6.1
requests==2.31.0 